Move proto related tests to own file
